+++ Firestore/core/test/unit/core/pipeline/collection_test.cc
@@ -223,8 +223,8 @@ TEST_F(CollectionTest, WhereOnValues) {
 
 TEST_F(CollectionTest, WhereInequalityOnValues) {
   RealtimePipeline pipeline = StartPipeline("/users");  // Use RealtimePipeline
-  auto where_expr =
-      GtExpr({std::make_shared<Field>("score"), SharedConstant(80LL)});
+  auto where_expr = GtExpr({std::make_shared<Field>("score"),
+                            SharedConstant(static_cast<int64_t>(80LL))});
   pipeline = pipeline.AddingStage(std::make_shared<Where>(where_expr));
 
   auto doc1 = Doc("users/bob", 1000, Map("score", 90LL));
@@ -238,8 +238,8 @@ TEST_F(CollectionTest, WhereInequalityOnValues) {
 
 TEST_F(CollectionTest, WhereNotEqualOnValues) {
   RealtimePipeline pipeline = StartPipeline("/users");  // Use RealtimePipeline
-  auto where_expr =
-      NeqExpr({std::make_shared<Field>("score"), SharedConstant(50LL)});
+  auto where_expr = NeqExpr({std::make_shared<Field>("score"),
+                             SharedConstant(static_cast<int64_t>(50LL))});
   pipeline = pipeline.AddingStage(std::make_shared<Where>(where_expr));
 
   auto doc1 = Doc("users/bob", 1000, Map("score", 90LL));
